mega_downloader: Fallback temporary solution
Temporary fix for new link (mega,nz/file) etc.

+++ userbot/modules/mega_downloader.py
@@ -18,15 +18,17 @@
 #  OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 #  ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from subprocess import PIPE, Popen
+from asyncio import create_subprocess_shell as asyncSubprocess
+from asyncio.subprocess import PIPE as asyncPIPE
 
+import asyncio
 import re
 import json
 import os
-import time
+import multiprocessing
+import errno
 
 from pySmartDL import SmartDL
-from os.path import exists
 from urllib.error import HTTPError
 
 from userbot import CMD_HELP, LOGS
@@ -34,53 +36,48 @@ from userbot.events import register
 from userbot.modules.upload_download import humanbytes
 
 
-async def subprocess_run(cmd, megadl):
-    subproc = Popen(cmd, stdout=PIPE, stderr=PIPE,
-                    shell=True, universal_newlines=True)
-    talk = subproc.communicate()
+async def subprocess_run(megadl, cmd):
+    subproc = await asyncSubprocess(cmd, stdout=asyncPIPE, stderr=asyncPIPE)
+    stdout, stderr = await subproc.communicate()
     exitCode = subproc.returncode
     if exitCode != 0:
         await megadl.edit(
-            '```An error was detected while running the subprocess:\n'
-            f'exit code: {exitCode}\n'
-            f'stdout: {talk[0]}\n'
-            f'stderr: {talk[1]}```')
-        return
-    return talk
+            '**An error was detected while running subprocess**\n'
+            f'```exitCode: {exitCode}\n'
+            f'stdout: {stdout.decode().strip()}\n'
+            f'stderr: {stderr.decode().strip()}```')
+        return exitCode
+    return stdout.decode().strip(), stderr.decode().strip(), exitCode
 
 
 @register(outgoing=True, pattern=r"^.mega(?: |$)(.*)")
 async def mega_downloader(megadl):
     await megadl.edit("`Processing...`")
-    textx = await megadl.get_reply_message()
+    msg_link = await megadl.get_reply_message()
     link = megadl.pattern_match.group(1)
     if link:
         pass
-    elif textx:
-        link = textx.text
+    elif msg_link:
+        link = msg_link.text
     else:
-        await megadl.edit("`Usage: .mega <mega url>`")
-        return
-    if not link:
-        await megadl.edit("`No MEGA.nz link found!`")
-    await mega_download(link, megadl)
-
-
-async def mega_download(url, megadl):
+        return await megadl.edit("Usage: `.mega <MEGA.nz link>`")
     try:
-        link = re.findall(r'\bhttps?://.*mega.*\.nz\S+', url)[0]
+        link = re.findall(r'\bhttps?://.*mega.*\.nz\S+', link)[0]
+        """ - Mega changed their URL again - """
+        if "file" in link:
+            link = link.replace("#", "!").replace("file/", "#!")
+        elif "folder" in link or "#F" in link or "#N" in link:
+            await megadl.edit("`Currently support folder download are removed`.")
+            return
     except IndexError:
-        await megadl.edit("`No MEGA.nz link found`\n")
-        return
+        return await megadl.edit("`No MEGA.nz link found`\n")
     cmd = f'bin/megadown -q -m {link}'
-    result = await subprocess_run(cmd, megadl)
+    result = await subprocess_run(megadl, cmd)
     try:
         data = json.loads(result[0])
     except json.JSONDecodeError:
-        await megadl.edit("`Error: Can't extract the link`\n")
-        return
-    except TypeError as e:  # in case file exists log to heroku then return
-        LOGS.info(str(e))
+        return await megadl.edit("`Error: Can't extract the link`\n")
+    except (IndexError, TypeError):
         return
     file_name = data["file_name"]
     file_url = data["url"]
@@ -94,9 +91,7 @@ async def mega_download(url, megadl):
     try:
         downloader.start(blocking=False)
     except HTTPError as e:
-        await megadl.edit("`" + str(e) + "`")
-        LOGS.info(str(e))
-        return
+        return await megadl.edit("`" + str(e) + "`")
     while not downloader.isFinished():
         status = downloader.get_status().capitalize()
         total_length = downloader.filesize if downloader.filesize else None
@@ -107,50 +102,61 @@ async def mega_download(url, megadl):
         estimated_total_time = downloader.get_eta(human=True)
         try:
             current_message = (
-                f"**{status}**..."
-                f"\nFile Name: `{file_name}`\n"
-                f"\n{progress} `{percentage}%`"
+                "File Name:"
+                f"\n`{file_name}`\n\n"
+                "Status:"
+                f"\n**{status}** | {progress} `{percentage}%`"
                 f"\n{humanbytes(downloaded)} of {humanbytes(total_length)}"
                 f" @ {speed}"
                 f"\nETA: {estimated_total_time}"
             )
-            if status == "Downloading":
+            if display_message != current_message:
                 await megadl.edit(current_message)
-                time.sleep(0.2)
-            elif status == "Combining":
-                if display_message != current_message:
-                    await megadl.edit(current_message)
-                    display_message = current_message
-        except Exception as e:
-            LOGS.info(str(e))
+                await asyncio.sleep(0.2)
+                display_message = current_message
+        except Exception:
+            pass
+        finally:
+            if status == "Combining":
+                await asyncio.sleep(float(downloader.get_eta()))
     if downloader.isSuccessful():
         download_time = downloader.get_dl_time(human=True)
-        if exists(temp_file_name):
-            await decrypt_file(
-                file_name, temp_file_name, hex_key, hex_raw_key, megadl)
-            await megadl.edit(f"`{file_name}`\n\n"
-                              "Successfully downloaded\n"
-                              f"Download took: {download_time}")
+        try:
+            P = multiprocessing.Process(target=await decrypt_file(megadl,
+                                        file_name, temp_file_name,
+                                        hex_key, hex_raw_key),
+                                        name="Decrypt_File")
+            P.start()
+            P.join()
+        except FileNotFoundError as e:
+            return await megadl.edit(str(e))
+        else:
+            return await megadl.edit(f"`{file_name}`\n\n"
+                                     "Successfully downloaded\n"
+                                     f"Download took: {download_time}")
     else:
-        await megadl.edit("Failed to download...")
+        await megadl.edit("`Failed to download, "
+                          "check heroku Logs for more details`")
         for e in downloader.get_errors():
             LOGS.info(str(e))
     return
 
 
-async def decrypt_file(file_name, temp_file_name, hex_key, hex_raw_key, megadl):
-    await megadl.edit("Decrypting file...")
+async def decrypt_file(megadl, file_name, temp_file_name,
+                       hex_key, hex_raw_key):
     cmd = ("cat '{}' | openssl enc -d -aes-128-ctr -K {} -iv {} > '{}'"
            .format(temp_file_name, hex_key, hex_raw_key, file_name))
-    await subprocess_run(cmd, megadl)
-    os.remove(temp_file_name)
+    if await subprocess_run(megadl, cmd):
+        os.remove(temp_file_name)
+    else:
+        raise FileNotFoundError(
+            errno.ENOENT, os.strerror(errno.ENOENT), file_name)
     return
 
 
 CMD_HELP.update({
     "mega":
-    ".mega <mega url>\n"
-    "Usage: Reply to a mega link or paste your mega link to\n"
-    "download the file into your userbot server\n\n"
-    "Only support for **FILE** only.\n"
+    ">`.mega <MEGA.nz link>`"
+    "\nUsage: Reply to a MEGA.nz link or paste your MEGA.nz link to "
+    "download the file into your userbot server."
 })
